refactor main giant class into multiple traits
