+++ src/Services/OdkLinkServices/OdkProjectService.php
@@ -1,0 +1,109 @@
+<?php
+
+namespace Stats4sd\FilamentOdkLink\Services\OdkLinkServices;
+
+
+use Illuminate\Http\Client\ConnectionException;
+use Illuminate\Http\Client\RequestException;
+use Illuminate\Support\Facades\Http;
+use Illuminate\Support\Str;
+use Stats4sd\FilamentOdkLink\Models\OdkLink\OdkProject;
+
+
+trait OdkProjectService
+{
+
+    /**
+     * Creates a new project in ODK Central
+     *
+     * @return array $projectInfo
+     *
+     * @throws RequestException|ConnectionException
+     */
+    public function createProject(string $name): array
+    {
+        $token = $this->authenticate();
+
+        // prepend platform identifier to project name;
+        $name = (config('app.short_name') ?? config('app.name')) . '- ' . $name;
+
+        // leave 7 characters for the "all " prefix and number suffix for the app user;
+        if (Str::length($name) > 57) {
+            $name = Str::squish($name);
+        }
+
+        if (Str::length($name) > 57) {
+            $name = Str::limit($name, limit: 57, end: '');
+        }
+
+        return Http::withToken($token)
+            ->post("{$this->endpoint}/projects", [
+                'name' => $name,
+            ])
+            ->throw()
+            ->json();
+    }
+
+    public function createProjectAppUser(OdkProject $odkProject): array
+    {
+        $token = $this->authenticate();
+
+        // truncate name to 64 characters
+        $displayName = Str::limit('All ' . $odkProject->name . ' ' . $odkProject->appUsers()->count() + 1, limit: 64, end: '');
+
+        // create new app-user
+        $userResponse = Http::withToken($token)
+            ->post("{$this->endpoint}/projects/{$odkProject->id}/app-users", [
+                'displayName' => $displayName,
+            ])
+            ->throw()
+            ->json();
+
+        // assign user to all the forms in the project
+        Http::withToken($token)
+            ->post("{$this->endpoint}/projects/{$odkProject->id}/assignments/manager/{$userResponse['id']}")
+            ->throw()
+            ->json();
+
+        return $userResponse;
+    }
+
+    /**
+     * Updates a project name
+     *
+     * @return array $projectInfo
+     *
+     * @throws RequestException|ConnectionException
+     */
+    public function updateProject(OdkProject $odkProject, string $newName): array
+    {
+        $token = $this->authenticate();
+
+        return Http::withToken($token)
+            ->post("{$this->endpoint}/projects/$odkProject->id", [
+                'name' => $newName,
+            ])
+            ->throw()
+            ->json();
+    }
+
+    /**
+     * Archives a project
+     *
+     * @return array $success
+     *
+     * @throws RequestException|ConnectionException
+     */
+    public function archiveProject(OdkProject $odkProject): array
+    {
+        $token = $this->authenticate();
+
+        return Http::withToken($token)
+            ->post("{$this->endpoint}/projects/$odkProject->id", [
+                'name' => $odkProject->name,
+                'archived' => true,
+            ])
+            ->throw()
+            ->json();
+    }
+}
